Fix "needs sanitising" functions spilling over
When a function call to a function in the list of those that are
specifically marked as needing sanitization is nested within a call to
another function, any later parameters of the super-function will be
marked ass needing sanitization under some circumstances.

This is fixed by marking the end of the statement as the closing
parenthesis of the “needs sanitizing” function, rather than the next
semi-colon.