XSS sniff: Remove non-output functions from list
This list was added in 722b2c660f9bdd388afb622b743db883af4f1726, as
part of work on #33. The intention was to properly handle functions
like `_e()` which generate direct output. However, for some reason we
went overboard and added a bunch of l10n functions, most of which do
not print output.

I’m removing these non-printing functions.