Use tokenizer instead of regex in yoda conditions sniff
This fixes issues with comparison to constants not being flagged, and
probably will be less buggy overall.

There is one side-effect:

```php
// Before this was good:
if ( $true == strtolower( $check ) ) {}

// Now it’s bad, but this is OK:
if ( strtolower( $check ) == $true  ) {}
```

I think this is the correct behavior.

Another thing to consider is when comparing a variable to a variable. I
don’t remember what the previous behavior was, but this is current
behavior:

```php
if ( $foo == $bar ) {} // Bad
if ( ( $foo ) == $bar ) {} // OK
```

Finally, the matching code is no longer included in the error message.
It’s possible to add that back in, but it would take some work.

Feedback welcome. :-)

Fixes #290