Remove redundant sanitization check
This redundant check for sanitization was introduced sometime a while
ago. It is actually a bug, because it makes it so that no validation
check is performed if sanitization isn’t applied.

This is fixed by removing it, as demonstrated by the fact that we now
get two errors on line 33, instead of just the one for missing
sanitization.

I’ve added another test on line 85, for a sanitized but not validated
variable, just to demonstrate that it works correctly. This behavior is
unchanged; an error would have been given here previously.