Separate sanitizing and escaping functions
Historically there has a been a list of “sanitizing” functions in the
XSS sniff. This list could be added to via the
`customSanitizingFunctions` property. This list of functions was also
used by the `ValidatedSanitizedInput` sniff.

However, what the XSS sniff is really checking for is that the output
is *escaped*. This is slightly different from *sanitization*, and is
(usually) performed by different functions.

In this commit, I’ve split the `sanitizingFunctions` list in the XSS
sniff, and moved it to the main sniff. This will allow the list to be
accessed from all sniffs more easily. I’ve also flipped the list so
that the functions are the keys instead of values, because this allows
for quicker lookup with `isset()` instead of having to use
`in_array()`. The values currently have no meaning, though it’s
possible we’ll use them for something in the future. (I’ve also done
this for the other function lists from the XSS sniff, and the list of
nonce verification functions.)

I’ve split the list of “sanitizing” functions into the
`sanitizingFunctions` and `escapingFunctions` properties. A few
functions are found on both lists, but for the most part there is
pretty clear separation. We will tweak this in the future as needed.

I’ve completely removed the `balanceTags()` function from both lists,
because it doesn’t actually strip scripts or anything like that. There
may be a case for adding it back to the sanitizing functions list, and
you’re welcome to make it.

I’ve also moved `ent2ncr()` to the list of formatting functions, since
it doesn't actually escape things, just format them. (It was added in
#189.)

The `customSanitizingFunctions` property of the XSS sniff has been
deprecated, and a `customEscapingFunctions` property has been added in
its place. When the deprecated property is used, a warning will be
given.

I’ve also added a `customSanitizingFunctions` property to the
`ValidatedSanitizedInput` sniff. This is where custom sanitizing
functions should be added now. Adding them to either of the properties
of the XSS sniff won’t have the desired effect.

Finally, there is one issue yet to be resolved, so the tests are
intentionally failing. There is currently no provision made for code
like this, which takes user input and directly escapes it for display:

```php
echo esc_html( $_POST[‘foo’] );
```

With a little work, we could make this permissible. But perhaps there
is a case to be made that sanitization should still be applied even in
a case like this. Please let your opinion be known.

Fixes #343