Remove old code for preventing dup messages
This code was added long ago
(aa57302aa71b8afd1aa2de4e6ec51498df858315), but since then the sniff
has evolved to the point that it is no longer needed. Specifically, the
sniff is much more restrictive about when it flags `$wpdb`, limiting
the errors to a predefined list of methods. One reason that duplicate
messages used to be given on the same line was when you called
`$wpdb->query( $wpdb->prepare( /* … */ ) );`, and both methods was be
flagged. Because the `prepare()` method is no longer flagged, this
problem no longer occurs.

For good measure, I’ve also modified the sniff to return the
`$endOfStatment` pointer, which will tell PHPCS not to sniff any more
code until after the semicolon at the end of the statement. This would
also prevent duplicate errors, and may even provide a small performance
benefit as well.