Check spacing of functions
This adds functions and closures to the control structure spacing
sniff. The sniff checks for proper spacing on each side of each of the
parentheses. It also checks that the opening brace is on the same line.

I’ve actually improved the sniff while I was at it, to check for excess
space in addition to insufficient space.

Because of these improvements, and due to the non-configurability of
the upstream version of this sniff, this sniff is really not a
duplicate, and so #385 can be closed.

Fixes #412, #385