XSS sniff: use correct token in error messages
Previously, the incorrect token would be used when the error message
was for a function. This is because the `$i` variable is fast-forwarded
to skip over all or part of the function call’s arguments. The result
was that the error would be given for the closing parenthesis of the
function, instead of the function name. This is now fixed.

Discovered while working on #442.