Fix false positive for yoda conditions
When a string was on the left hand side of the condition and something
beside a variable was on the right hand side, the condition would still
be flagged.

To fix that we ignore a condition when we find a string on the left
hand side. We do have to add a check to make sure that the string isn’t
the key of an array assignment (`$var[‘key’]`). We do that by checking
for a closing square bracket, and assuming the condition is an
assignment when we find one.

Fixes #509