Introduce new abstract class WordPress_AbstractFunctionParameterSniff
This extends the `WordPress_AbstractFunctionRestrictionsSniff` class for the yes/no function call determination, retrieves the parameters used in a function call and then passes processing on to the child class.

The child class is expected to implement a `process_parameters()` method to handle the actual testing of parameters.

Optionally a `process_no_parameters()` method can be added to the child class to handle cases where parameters are expected, but not found.

The abstract expects a `$target_functions` array, but is without opinion about how that array should look, except for the requirement that the top level array keys are the function names.
Further requirements of the array are not needed as the processing of the parameters is done in the child class, allowing for great flexibility in potential use of this class.