FileNameUnitTest: Re-enable the unit tests for non-strict class names.
The unit tests for non-strict class names were never enabled as the directory in which the test files live, wasn't listed in the `glob()` call.
This was previously not corrected as the `@codingStandardsChangeSetting` directives did not seem to work well in combination with a PHP open tag as the target token.

This has now been fixed. The unit tests should now pass and the directory has been included by making the `glob()` call more dynamic which will automatically include additional directories if these would be added in the future.