:sparkles: New sniff to verify correct array indentation
* Multi-line arrays should be indented by one tab from the beginning of the line containing the array opener.
* The array closer should be level with the indentation of the line containing the array opener.

While the `ScopeIndent` sniff checks this type of alignment for most structures, it turns out it doesn't do so for arrays (I suspect this is because PHPCS itself prefers a different array style).

This PR fixes that by adding a new sniff. Both the errors this sniff throws have been made fixable.

The new sniff has been added to the `WordPress-Core` ruleset.