:sparkles: New sniff: FunctionCallSignatureNoParams
Report on and remove whitespace between function call parenthesis when the function call has no parameters.

This sniff *will* result in contradictory messages when run with PHPCS, but that can't really be helped.

```
 5 | ERROR | [x] Expected 1 spaces after opening bracket; 6 found
   |       |     (PEAR.Functions.FunctionCallSignature.SpaceAfterOpenBracket)
 5 | ERROR | [x] Expected 1 spaces before closing bracket; 6 found
   |       |     (PEAR.Functions.FunctionCallSignature.SpaceBeforeCloseBracket)
 5 | ERROR | [x] Function calls without parameters should have no spaces
   |       |     between the parenthesis.
   |       |     (WordPress.Functions.FunctionCallSignatureNoParams.WhitespaceFound)
```

Includes an auto-fixer.

This auto-fixer has been tested to not conflict with the `PEAR.Functions.FunctionCallSignature` sniff.

Fixes issue 970