:sparkles: New `CodeAnalysis.EmptyStatement` sniff
New sniff to detect superfluous semi-colons and empty PHP open-close tag combinations.

Includes an auto-fixer.

Added into a new `CodeAnalysis` category to be in line with the related upstream sniff.

As these issues are not explicitly mentioned in the WP Core Coding Style Handbook and are more along the lines of best practices, I've added the sniff to the `WordPress-Extra` ruleset.
For that same reason, I've elected to let the sniff throw warnings instead of errors.

For the current WP Core codestyle update, it would, however, be advisable to explicitly enable this sniff from the ruleset as issues which this sniff will fix, *have* already been detected in WP Core during a visual inspection.

Fixes 975

N.B. This sniff should eventually be merged with the upstream `Generic.CodeAnalysis.EmptyStatement` sniff and pulled there. This would fix upstream issue #1119.