ArrayIndentation sniff: Fix disappearing close brace when close brace not on its own line
For multi-line array declarations, the close brace of the array is supposed to be on its own line + aligned with the line containing the array keyword.

The `ArrayDeclaration` sniff will move the close brace to its own line.
The `ArrayIndentation` sniff is only intended to fix the close brace indentation.

The `ArrayIndentation` sniff was until now - incorrectly - presuming the close brace would already be on its own line with only whitespace before it.

While the array *item* indentation part of the sniff did not suffer from this issue, I've build in a little extra check in that part just to make sure.