PreparedSQLPlaceholders: throw warning on replacements/no placeholders
If no (valid) placeholders are found, but replacement variables are passed, this indicates either an unfinished query or a query using invalid placeholders, so throw a warning.

Also make sure that a valid escaped % literal followed by a placeholder is properly counted as a one placeholder so the message won't throw a false positive.