:sparkles: New ArbitraryParenthesesSpacing sniff for Core
Checks the spacing on the inside of arbitrary parentheses.

Arbitrary parentheses, for the purposes of this sniff, are all parentheses which do not belong to a control structure, function declaration, function call or language construct.
This is to prevent conflicts and duplication with rules for those type of constructs.

The sniff will throw (fixable) errors when an incorrect amount of whitespace (too much/too little) has been found on the inside or the parenthesis.
The sniff will throw a warning when an empty set of arbitrary parentheses is encountered.

Includes extensive unit tests.

The sniff has been set up to be pulled upstream and once upstream has been merged and the minimum PHPCS requirement for WPCS has gone up, to be removed from WPCS.

Fixes 1194

- [ ] Add info on the custom properties to the wiki