MultipleStatementAlignment: property name consistency
Everywhere else where a "ignore new lines" property is used, both in WPCS as well as in PHPCS upstream, the property is called `ignoreNewlines`.

Only just now realized that the `s` was missing from the property in the new `Arrays.MultipleStatementAlignment` sniff.

This fixes that.

Not a BC-break as the sniff has now been in a release yet.