Add a couple of extra sniffs to `Extra`
A couple of sniffs which were previously in VIP are also useful for `Extra`. The messages in these sniffs have - where appropriate - been downgraded to warnings.

This commit adds these sniffs to the `Extra` ruleset.

Note: There are three more candidates for this:
* `DB.DirectDatabaseQuery`
* `DB.SlowDBQuery`
* `Security.ValidatedSanitizedInput`

I've not included the first two (yet) as those may need some more discussion first.
The third one is not included as it throws a lot of noise messages and has quite some open issues surrounding it, which should be addressed first IMO.