Fix build failure
The build failure has to do with an upstream fixer conflict between two sniffs fighting over tabs vs spaces.
This is related to a recent upstream change and should be reported there for further attention.

As the space is not important for this particular unit test, let's just prevent the conflict from happening for now.