PrecisionAlignment: fix bug - examine inline HTML at end of of file
The new line character is normally tokenized as `T_WHITESPACE`, however, not so when at the end of inline HTML, where it is tokenized as part of the `T_INLINE_HTML` token.

Bowing out of the sniff when no next token could be found, prevent the last line of inline HTML at the end of a file from being examined.