AlternativeFunctions: only advise alternative if min WP version supports it
Only show suggestions for WP functions which are recommended as alternative for native PHP functions, when the `minimum_supported_version` contains the recommended WP function.

Includes unit test.

Note: the unit test file now contains a property setting. This property needs to be reset after use. The reset value has to be in-line with the default value as set in the `Sniff` class.
A note to that end has been added to the property documentation in the `Sniff` class.