Composer: various tweaks
This commit adds two additional scripts and adjusts the existing one.

The additional scripts are just shortcuts to run `phpcs` and `phpcbf`.

The more important change is how the scripts are run - using the `@php` prefix ensures that the scripts are run against the same PHP version as with which Composer is called, instead of against the system default PHP version.

This is important if - as a WPCS dev - you change the PHP version for Composer to be able to test things against different PHP versions.

Ref: https://github.com/composer/composer/issues/7645