Sniff::get_declared_namespace_name(): improve code-style independence
A namespace declaration can contain whitespace and comments and PHP will just ignore the whitespace and comments when reading the code. Also see: squizlabs/PHP_CodeSniffer 2150

While it is rare to encounter such code, this utility method should handle it correctly.

As for the `$namespaceName` to be returned by the method, this should not contain any whitespace or comments encountered. That way the namespace name will be "clean" for examination by sniffs.

Includes unit test via the `GlobalVariablesOverride` sniff.