I18nTextDomainFixer: improve handling of whitespace in headers
While WPCS does not allow for inline tab alignment, this sniff should respect the tab vs space choice made in the original code.

This was already handled correctly for the function call parts of the sniff.

This commit improves this for the plugin/theme header part of the sniff.

Includes unit tests.