PrefixAllGlobals: improve error position precision
... for hook names and constant names when declared using `define()`.

The sniff would throw the error message on the line which contained the first token regarded as part of the parameter - which might be end of line whitespace on the line before the actual hook name or constant name  in a multi-line function call.

This minor change improves the line number precision for these situations.

Fixed 1495