Sniff::merge_custom_array(): Remove support for incorrectly set custom array properties
Array properties passed in from a custom ruleset need to have a `type="array"` attribute.

Until now WPCS was lenient when this attribute was missing and would convert an array passed as a comma delimited string to an array.

This functionality was deprecated for the only property which still "officially" allowed for a comma-delimited string list in WPCS 1.0.0 and will now be removed.

The only exception are the two array properties which can also be set via the command-line, i.e. the `prefixes` property for the `WordPress.NamingConventions.PrefixAllGlobals` sniff and the `text_domain` property for the `WordPress.WP.I18n` sniff.
If - and only if - these properties are set from the command-line, the sniffs will still convert the received string to an array.

Refs:
* 1390

This also removes the leniency for spaces surrounding array values.

PHPCS 3.2.0 includes a similar fix as was previously made in PR 1185, which means that properties passed as arrays will already be trimmed, so no need to do it again.

Refs:
* 1185
* squizlabs/PHP_CodeSniffer 1689