Sniff::$unslashingSanitizingFunctions: add doubleval() and count()
While `doubleval()` is an alias of `floatval()` and shouldn't be used, for the purposes of the ValidatedSanitizedInput sniff, both functions should be recognized.

And as `count()` doesn't actually access the data in the variable, but only counts the number of elements, it is also safe to use without unslashing/sanitizing the variable beforehand.
Same goes for the `sizeof()` alias of `count()`.

Includes unit tests.

Fixes 1659