ValidatedSanitizedInput: allow for validation using key_exists()
This builds onto the previous enhancement made in 1635 which started recognizing `array_key_exists()` as a way to validate a variable.

`key_exists()` is an alias for `array_key_exists()` and while alias functions shouldn't be used,  for the purposes of the ValidatedSanitizedInput sniff, both functions should be recognized.

Includes unit test.

Notes:
* Removes the `array_key_exists()` method from the list of `$sanitizingFunctions` as it doesn't belong there and is now handled differently anyway (this should have been removed in 1635).
* Updates the version numbers for the change in the method documentation. We never released version 2.0.1, so both this change as well as the one from 1635 will now be released in 2.1.0.