:sparkles: New CodeAnalysis/EscapedNotTranslated sniff
At times, people will accidentally forget to add the `__` when they intend to use one of the "translate + escape" functions.
I've run into this a number of times now when reviewing/fixing code and found that they aren't that easy to spot visually when you're focussed on other things.

So as it was such an easy sniff to write, I figured I may as well.

AFAICS there are only two escaping functions in core which have direct "translate + escape" sister-functions.
All the same, the sniff has been set up to allow for more similar function-combis to be added.

Includes unit tests.
Includes documentation.

Sniff has been added to the `WordPress-Extra` ruleset.