ArrayKeySpacingRestrictions: fix whitespace violations in one go
If the array key should not be surrounded by spaces and there was more than one whitespace token between a bracket and the array key, the fixer would take several loops to remove the consecutive whitespace tokens.

This has been changed to fixing the whitespace in one go, reducing the chance of fixer conflicts and making the sniff more efficient.

Includes unit tests.

This can be tested by running the sniff with the `-v` option. Without this fix, the sniff takes 4 loops when fixing the unit test cases in this commit, with the fix, it takes 2 loops.