ValidFunctionName: use PHPCSUtils `isMagicMethodName()`
The magic method name list in PHPCSUtils is more complete than the one in the PEAR sniff.

Note: As this sniff bows out for methods declared in extended classes anyway, the PHP native double underscore methods from the SOAPClient class do not need to be taken into account, which is why the `isMagicMethodName()` method is used instead of the `isSpecialMethodName()` method.

Refs:
* [isMagicMethodName()](https://phpcsutils.com/phpdoc/classes/PHPCSUtils-Utils-FunctionDeclarations.html#method_isMagicMethodName)
* [isSpecialMethodName()](https://phpcsutils.com/phpdoc/classes/PHPCSUtils-Utils-FunctionDeclarations.html#method_isSpecialMethodName)