NamingConventions/ValidFunctionName: implement PHPCSUtils methods
PHP is largely case-insensitive for class and method name, but not completely: the case-insensitivety does not apply to the non-ascii range.

The PHPCSUtils methods recognize this correctly.

Includes minor efficiency tweak by only doing the PHP4 constructor/destructor comparison when it's a real class as a name match can never be made for anonymous classes anyway.

Includes unit tests.

Note: the test would previously pass, but only due to the `strtolower()` mangling both the class name as well as the method name equally.