EnqueuedResources: bug fix - handling of multi-line text strings
As it was, if the text the sniff is looking for was split over multiple lines, the sniff would incorrectly not throw an error (false negative).

By using the PHPCSUtils `TextStrings::getCompleteTextString()` method, we can get access to the complete content of a multi-line text string and prevent the false negative.

This does create two minor complications though:
1. The complete text string may contain multiple link/script tags, so we need to use `preg_match_all()` instead of `preg_match()` to be able to report on each of the matches.
2. We need to figure out on which line of the multi-line text string the matched content occurred so we can throw the error on the correct line.
    Luckily, as multi-line text strings have only one token per line, that's easy enough to work out, especially if we let `preg_match_all()` remember the offsets at which a match was found, so we also won't confuse one match for another.

Includes unit tests.

Co-authored-by: Juliette <663378+jrfnl@users.noreply.github.com>