Add Code of Conduct
Based on the [Contributor Covenant](https://www.contributor-covenant.org) version 2.0.

As there is no email address available for the project, I'm proposing to use the WP Core Slack channel #core-coding-standards as the official channel to report violations. Depending on the violation, details can shared in private via DMs if needed to protect the privacy of the parties involved.