Core: move last modifier keyword sniff from Extra to Core
While this was not explicitly mentioned in the Make post, it is implied that there should be exactly one space after each modifier keyword for properties and methods.

This was already being checked for in `Extra`. I'm proposing to move this sniff to the `Core` ruleset.

Note: I've checked and introducing this sniff to the `Core` ruleset does not yield any issues for WP Core as is, so it would be a silent introduction.

Refs:
* WordPress/WordPress-Coding-Standards 1101