Update the minimum_wp_version to WP 5.8
The minimum version should be three versions behind the latest WP release, so what with 6.1 being out, it should now be 5.8.

Includes adding a little more variation to the example code in the property docblock in the `MinimumWPVersionTrait`.

Includes updating the list of deprecated functions in the `WP.DeprecatedFunctions` sniff based on the WP 6.1.1 release.
Input for this based on @JDGrimes's WP deprecated code scanner, though that is getting more and more difficult to run as it is no longer maintained and is throwing lots of errors.

Notes:
* The `sanitize_url()` function has been _un_deprecated in WP 5.9. See 2031.
* The replacement for the deprecated `wp_cache_reset()` function has been changed in WP Core.
    Ref: https://core.trac.wordpress.org/changeset/52705
* The `_filter_query_attachment_filenames()` function lists the wrong deprecation version in Core. WPCS now contains the correct version.
    A patch has been submitted to fix the listing in Core.
    Refs:
    - https://core.trac.wordpress.org/changeset/54524
    - https://core.trac.wordpress.org/ticket/56791#comment:10
* Other than the above, no verification has been executed on the newly added deprecated function information.

Includes minor tweak to make the generation of the test lines array a little more descriptive.

Previous: 1881

Closes 2031