WP/EnqueuedResources: minor code performance tweak
Always retrieve the stack pointer to the last token of a potentially multi-line text string and skip to the end after analyzing the content.

While the `TextStrings::getEndOfCompleteTextString()` and the `TextStrings::getCompleteTextString()` have some overlap, the results of the overlapping code is cached, so this should barely affect performance, while skipping to the end of the text string will greatly improve performance for files containing long multi-line text strings.