SnakeCaseHelper::get_suggestion(): bug fix - underscore handling
The `SnakeCaseHelper::get_suggestion()` method should have no opinion on the use of leading/trailing/double underscores in a name.

In practice, however, it did and the suggested name would be incorrect.

Fixed now.

Includes tests to safeguard this fix in the `ValidFunctionName` test case file, though the current methodology for testing would not catch a regression.

To illustrate the issue - originally, the error messages for these new tests would read:
```
216 | ERROR | Method name "___Leading_Underscores" in class UnderScoreHandling is not in snake case format, try "leading_underscores"
 217 | ERROR | Method name "Multiple_______Underscores" in class UnderScoreHandling is not in snake case format, try "multiple____underscores"
 218 | ERROR | Method name "Trailing_Underscores___" in class UnderScoreHandling is not in snake case format, try "trailing_underscores"
```

With the fix, the error messages for the new tests read:
```
 216 | ERROR | Method name "___Leading_Underscores" in class UnderScoreHandling is not in snake case format, try "___leading_underscores"
 217 | ERROR | Method name "Multiple_______Underscores" in class UnderScoreHandling is not in snake case format, try "multiple_______underscores"
 218 | ERROR | Method name "Trailing_Underscores___" in class UnderScoreHandling is not in snake case format, try "trailing_underscores___"
```