PHPCS: ruleset tweak
For this repo, we use a separate package for linting code, so there is no need to have the `Generic.PHP.Syntax` sniff running when doing a PHPCS scan (the sniff is included in the `WordPress-Extra` ruleset).

An analysis of the performance of a sniff run on this repo revealed that this was the slowest sniff by far. As excluding it will not have any QA impact, let's make CS runs on the code in this repo faster.