Helpers/MinimumWPVersionTrait: reduce use of magic number
The default value for the `minimum_wp_version` property is used in a number of places, mostly in the test files when resetting a changed value for the property to its default, making it a "magic number".

Preferably, a magic number should be assigned to a constant and the constant should be used everywhere.
Unfortunately, that's not possible as a) constants in traits are not allowed prior to PHP 8.2 and b) code in test case files does not get executed, so using the constant is not an option anyway.

This commit now makes couple of changes to improve/stabilize working with this "magic" number.
* The magic number is now only set in one place  - in the new `$default_minimum_wp_version` property.
* A user-provided `minimum_wp_version` property value from a ruleset will now be validated in the same way as a CLI value for the same.
    If no value is found or an invalid value is found, the default value for the `minimum_wp_version` property will be used.
    Includes making it explicit in the `composer.json` file that the `Filter` extension is a prerequisite for this package.
* In test case files, a "reset" of the `minimum_wp_version` property can now unset the value, lowering the maintenance burden when the default value for the `minimum_wp_version` property changes.
    Note: test files for the `Deprecated*` sniffs will still need updating whenever the value changes.

---

I wonder if we should rename the `get_wp_version_from_cli()` method to a better name. The method was already renamed in WPCS 3.0.0, so renaming it again does not add a breaking change which wasn't already there in the first place.

The reason is that, on the one hand, the `get_` prefix suggests that the method returns a value, which it doesn't. And on the other hand, the `from_cli` isn't fully covering the functionality anymore.

Name suggestions welcome.

Some (rough) ideas:
* `validate_minimum_wp_version()`
* `set_minimum_wp_version()`

Opinions ?

---

Fixes 2171