DB/PreparedSQL: use predefined token collections [5]
Leverage predefined token collections in PHPCS and PHPCSUtils to fill the list of `$ignored_tokens`.

This commit start using the PHPCSUtils `incrementDecrementOperators()` collection.

This effectively adds two additional tokens to the "safe" list, `++` and `--`. These tokens, in and of itself, are not problematic, so I believe it is safe to include these.

It could even be argued that increment/decrement tokens should be treated the same as "safe casted" tokens, as incrementing or decrementing a variable effectively turns it into a benign integer.

Includes a test demonstrating the issue.